Added Graph For Viewing Group Evaluations For Assignments
I added to the models and controllers to get data from the database to construct the graph. At the moment the graph displays only data for group evaluations, not peer evaluations. The information is displayed by clicking the analytics button from selecting a class and then selecting an assignment. The graph is then populated with data and if you click a bar it will show you the evaluations received that made the average for the bar clicked on. At the moment the graph does not hide who gave what evaluation for students, although I can easily add it and it will be done before the next push. In the next push I hope to accomplish th peer evaluations graph as it will now only involve a lot of copying and pasting.