Added the peer evaluation graph
In this update, the peer evaluation graph was added. Also I fixed the graph from a basic users perspective to no longer show names of people who gave them evaluations, in their stead it gives an anonymous name. The peer evaluation graph is different from the group evaluation in that there are no groups in the peer evaluation graph. The students names are listed on the y axis and the average rating on the x. The sort option now has flip flopped use for the peer evaluation and group evaluation graph. For the peer evaluation graph you can sort the names for averages by highest and lowest score as well as alphabetical first name ascending and descending. This functionality is available in the base ratings graph for group projects. The base ratings for peer evaluations can be sorted by highest and lowest score while the group averaged evaluations graph can be sorted by the highest and lowest score as well.